fix(journey-maps): updated examples
