+++ src/components-examples/journey-maps/angular/index.ts
@@ -8,13 +8,12 @@ import { SbbNotificationModule } from '@sbb-esta/angular/notification';
 import { SbbRadioButtonModule } from '@sbb-esta/angular/radio-button';
 import { SbbSelectModule } from '@sbb-esta/angular/select';
 import { SbbTabsModule } from '@sbb-esta/angular/tabs';
-import { JourneyMapsRoutesPoisExample } from '@sbb-esta/components-examples/journey-maps/angular/journey-maps-routes-pois/journey-maps-routes-pois-example';
 import { SbbJourneyMapsModule } from '@sbb-esta/journey-maps';
 
 import { JourneyMapsBasicExample } from './journey-maps-basic/journey-maps-basic-example';
+import { JourneyMapsFeatureTypesExample } from './journey-maps-feature-types/journey-maps-feature-types-example';
 import { JourneyMapsMapNavigationExample } from './journey-maps-map-navigation/journey-maps-map-navigation-example';
 import { JourneyMapsPoisMarkersExample } from './journey-maps-pois-markers/journey-maps-pois-markers-example';
-import { JourneyMapsRoutesZonesExample } from './journey-maps-routes-zones/journey-maps-routes-zones-example';
 import { JourneyMapsStyleOptionsExample } from './journey-maps-style-options/journey-maps-style-options-example';
 import { JourneyMapsUiOptionsExample } from './journey-maps-ui-options/journey-maps-ui-options-example';
 
@@ -23,18 +22,16 @@ export {
   JourneyMapsStyleOptionsExample,
   JourneyMapsMapNavigationExample,
   JourneyMapsPoisMarkersExample,
-  JourneyMapsRoutesZonesExample,
   JourneyMapsBasicExample,
-  JourneyMapsRoutesPoisExample,
+  JourneyMapsFeatureTypesExample,
 };
 const EXAMPLES = [
   JourneyMapsUiOptionsExample,
   JourneyMapsStyleOptionsExample,
   JourneyMapsMapNavigationExample,
   JourneyMapsPoisMarkersExample,
-  JourneyMapsRoutesZonesExample,
   JourneyMapsBasicExample,
-  JourneyMapsRoutesPoisExample,
+  JourneyMapsFeatureTypesExample,
 ];
 
 @NgModule({
