+++ src/components-examples/journey-maps/angular/journey-maps-feature-types/journey-maps-feature-types-example.css
